User commit

+++ src/main/java/com/example/demo/model/User.java
@@ -16,7 +16,7 @@ public class User {
 
     public User() {}
 
-    public User(Long id, String name, String email, String password, String role) {
+    public User(int id, String name, String email, String password, String role) {
         this.id = id;
         this.name = name;
         this.email = email;
@@ -24,8 +24,8 @@ public class User {
         this.role = role;
     }
 
-    public Long getId() { return id; }
-    public void setId(Long id) { this.id = id; }
+    public int getId() { return id; }
+    public void setId(int id) { this.id = id; }
     public String getName() { return name; }
     public void setName(String name) { this.name = name; }
     public String getEmail() { return email; }
